requirements.txt
# StackFlow — core dependencies
# API server, graph execution, database, observability.
# Module-specific deps live in modules/<name>/requirements.txt

# API framework
fastapi==0.128.3
uvicorn[standard]==0.40.0
orjson==3.11.7

# Graph execution
langgraph==1.0.7
langgraph-checkpoint-postgres==3.0.4
langgraph-sdk==0.3.3
langchain-core==1.2.9

# Observability
langfuse==3.12.1

# Database
psycopg[binary,pool]==3.3.2

# Data models
pydantic==2.12.5
pydantic-settings==2.12.0

# Environment
python-dotenv==1.2.1

# HTTP clients
httpx==0.28.1
requests==2.32.5
aiohttp==3.13.3

# Logging & output
colorlog==6.10.1
rich==14.3.2

# Templating
Jinja2==3.1.6

# Config & utilities
PyYAML==6.0.3
tenacity==8.5.0
